WIP: handle array/array better for UnionInPlace
We don't want to make bitmaps for small objects, and we don't
really want to make new bitmaps in general in cases where we
don't have to.

Also noticed that the "only one container with this key"
logic was sort of broken, probably because of the shift from
picking the first of a group to be the "target" to having
UnionInPlace be a method called on the target, so changed that
logic a bit.