add some extra sanity-checks
I don't trust the new ops code, so I've added a sanity-check to the
fragment cleanup for fragment_internal_test, which verifies that
every fragment's on-disk state matches the in-memory state after
it's done.

In a couple of the roaring import benchmarks, which intentionally bypass
the ops log for test-performance reasons (as is, we bypass the ops log
*for the part that we are not benchmarking*), this requires a direct
explicit snapshot. The UnionInPlace test needs this also because
UnionInPlace doesn't create an ops log for its changes.