don't copy the entire roaring data to write the op log
We don't need to allocate buffer space for a copy of the roaring
data and copy it into that buffer to save a write call, and
when the roaring data is kilobytes, that actually has very large
alloc/gc costs. Instead just make a second write call in those
cases.

But also make sure it's still checksummed.