support multiple workers for snapshot queue
Trying to track performance costs of delays a bit, but also
reduce them. We now report on delays getting into the snapshot
queue, but do actually get in there even if it takes a while.

The number 2 for snapshot workers may seem low, but it turns
out that increasing snapshot workers increases the actual
amount of work that needs to be done. Snapshots are way
more expensive than ops log. As long as something is in the
queue, it's getting writes to its ops log instead of
being snapshotted. That makes it faster. So if we clear the
queue more often, total amount of snapshotting increases for
the same number of writes, which means that we spend more
time overall doing the snapshots. So performance actually
gets *worse*. But if we don't clear the queue enough, and
there's enough fragments to fill it, we start getting larger
lag on operations.

Workers decrease latency (if and only if you're hitting the
queue length), but also decrease throughput.