use code generation to instantiate op types for Bitmap/Container
There's some commonalities between different operations, and
it'd be nice to be able to get something a bit more specific
than an interface{} that, for instance, definitely contains
*some* kind of bitmap operation, or *some* kind of container
operation. I think.

So we create named types, and interfaces, and this becomes
repetitive so there's code generation. I'm not sure this is the
right course of action.