drop generic lookup, add type-specific lookup, provide specialization for easy cases
For the easy cases where we know in advance of a specific operation
(and operation type) we want, it's useful to have a way to express
this concisely -- and literal type switches on an interface type let
us not bother with reflect for those common cases.

So the idea is, functions which we can implement on top of the
minimal generic Bitmap interface can just be provided as generics.
We indicate in optypes_list.txt that "ImportRoaring" is an
UpdateBytes operation on a Bitmap. This generates a top level
func ImportRoaring(Bitmap, []byte) (bool, int64, Bitmap). It
checks whether its target (the first parameter) provides
an implementation of
	ImportRoaringUpdateBytes([]byte) (bool, int64, Bitmap)
If it does, we use that. Otherwise, we call
	func genericImportRoaring(Bitmap, []byte) (bool, int64, Bitmap)

So if we write genericImportRoaring in terms of the general
Bitmap API (just using PutContainer, probably?), we're done -- but
specific implementations can streamline it if they want to.

This isn't actually implemented yet, but it's getting there.