fix silly typos and oversights
So the idea now is if you want to look up an op type, there's
OpLookup{OpType}, which will use the object's OpLookup function
if it has one, and there's OpLookupDirect{OpType}, which just
uses reflect. And corresponding Op{Data}LookupGeneric and
LookupGenericDirect functions which do this in the general case.
I haven't quite convinced myself those are useful, or that
they're not, yet.