Cleanup close/open logic a bit.
We don't need to closeStorage() before openStorage() with generations
available, because openStorage will automatically close the old
storage. Butw we also don't *want* to, because we would like the
old generation to be kept alive until the new one is ready and
the bitmap's been converted over to it, in principle.