Rework logic a bit more for data apply/import and source tracking
Simplifying this a bit and sharing the common case of "there is
no data". applyStorage wants to remap existing storage to match
the provided data, while importStorage wants to replace any existing
storage with the provided data. But they both assume that you want
to overwrite the file with your current data if there's no data
available.